+++ internal/stack/orderedmap.go
@@ -1,3 +1,6 @@
+// Copyright 2024 Josh Tompkin
+// Licensed under the MIT License
+
 package stack
 
 // Pair contains a Key of type K and Value of type V.
@@ -12,12 +15,6 @@ type OrderedMap[K comparable, V any] struct {
 	Pairs   map[K]*Pair[K, V]
 	list    []*Pair[K, V]
 	Current int
-}
-
-// NewOrderedMap returns a pointer to an OrderedMap with initialized values that
-// is ready to use.
-func NewOrderedMap[K comparable, V any]() *OrderedMap[K, V] {
-	return &OrderedMap[K, V]{Pairs: make(map[K]*Pair[K, V], 16), list: make([]*Pair[K, V], 0, 16)}
 }
 
 // Set sets OrderedMap[key] = Pair{Key: key, Value: val} and panics if the key
@@ -55,4 +52,10 @@ func (om *OrderedMap[K, V]) Next() (pair *Pair[K, V]) {
 // Reset sets the value of OrderedMap.Current to 0.
 func (om *OrderedMap[K, V]) Reset() {
 	om.Current = 0
+}
+
+// NewOrderedMap returns a pointer to an OrderedMap with initialized values that
+// is ready to use.
+func NewOrderedMap[K comparable, V any]() *OrderedMap[K, V] {
+	return &OrderedMap[K, V]{Pairs: make(map[K]*Pair[K, V], 16), list: make([]*Pair[K, V], 0, 16)}
 }