+++ internal/stack/stack.go
@@ -1,3 +1,6 @@
+// Copyright 2024 Josh Tompkin
+// Licensed under the MIT License
+
 package stack
 
 import (
@@ -7,13 +10,12 @@ import (
 	"strings"
 )
 
-const Suffix string = "\n"
-
 // Stack contains a slice of float64 and methods to operate on that slice.
 type Stack struct {
 	Values     []float64
 	Stash      float64
 	displayFmt string
+	// Expandable signifies whether stack capacity can be increased or not.
 	Expandable bool
 }
 
@@ -38,6 +40,9 @@ func (stk *Stack) Push(f float64) error {
 // Display returns a string of all values in the stack according to
 // Stack.displayFmt
 func (stk *Stack) Display() string {
+	if stk.displayFmt == "" {
+		return ""
+	}
 	ss := make([]string, len(stk.Values))
 	for i, f := range stk.Values {
 		ss[i] = fmt.Sprint(f)
@@ -57,34 +62,35 @@ type StackOperator struct {
 	Words       map[string]string
 	Stack       *Stack
 	Interactive bool
-	Prompt      func() string
+	Prompt      func() (prompt string)
 	PrintBuf    []byte
 	formatters  map[byte]func(*StackOperator) string
-	notFound    func(string) error
+	// notFound should return nil if the StackOperator does not care about
+	// entering missing input or an error if it does.
+	notFound func(string) error
 }
 
-// ParseInput splits input into words and either begins defining a calculator
-// word or parses each word as a token. It stops parsing input if it defines a
-// word or receives an error during token parsing, and returns the string and
-// error recieved. If it parses the entire input without error, it will return
-// the last string received and nil.
+// ParseInput splits an input string into words and interprets each word as a
+// token. It stops executing tokens if it encounters an '=' or the execution of
+// the token returns an error, and returns that error. ParseInput fills PrintBuf
+// with the message returned by the execution of the last token.
 func (so *StackOperator) ParseInput(input string) (err error) {
 	input = strings.TrimSpace(input)
 	split := strings.Split(input, " ")
 	for i, token := range split {
-        var s string
+		var s string
 		if token == "=" {
-            s, err = so.DefWord(split[i+1:])
-            so.PrintBuf = []byte(s)
-            return err
+			s, err = so.DefWord(split[i+1:])
+			so.PrintBuf = []byte(s)
+			return err
 		}
 		s, err = so.parseToken(token)
-        so.PrintBuf = []byte(s)
+		so.PrintBuf = []byte(s)
 		if err != nil {
 			return err
 		}
 	}
-    return err
+	return err
 }
 
 // DefWord adds a word to StackOperator.Words with the key being def[0] and the
@@ -93,7 +99,7 @@ func (so *StackOperator) ParseInput(input string) (err error) {
 // and an empty string and an error if not.
 func (so *StackOperator) DefWord(def []string) (message string, err error) {
 	if len(def) == 0 {
-		return "", errors.New(fmt.Sprintf(`define word: "= example 2 2 +"; remove word: "= example"%s`, Suffix))
+		return "", errors.New(fmt.Sprintf("define word: = example 2 2 +; remove word: = example\n"))
 	}
 	noEmpty := make([]string, 0, len(def))
 	for _, s := range def {
@@ -105,22 +111,28 @@ func (so *StackOperator) DefWord(def []string) (message string, err error) {
 		return "", nil
 	}
 	word := noEmpty[0]
-	if strings.Contains("0123456789=.", string(word[0])) {
-		return "", errors.New(fmt.Sprintf("could not define %s : cannot start word with digit, '=', or '.'%s", word, Suffix))
+	if _, err := strconv.ParseFloat(word, 64); err == nil {
+		return "", errors.New(fmt.Sprintf("counld not define %s : cannot redifine number\n", word))
+	}
+	forbidden := []string{"=", "quit"}
+	for _, s := range forbidden {
+		if word == s {
+			return "", errors.New(fmt.Sprintf("could not define %s : word cannot be any of: %s\n", word, strings.Join(forbidden, " ")))
+		}
 	}
 	if _, present := so.Actions.Get(word); present {
-		return "", errors.New(fmt.Sprintf("could not define %s : cannot redifine operator%s", word, Suffix))
+		return "", errors.New(fmt.Sprintf("could not define %s : cannot redifine operator\n", word))
 	}
 	if len(noEmpty) == 1 {
 		if _, present := so.Words[word]; !present {
-			return "", errors.New(fmt.Sprintf("could not delete %s : not defined%s", word, Suffix))
+			return "", errors.New(fmt.Sprintf("could not delete %s : not defined\n", word))
 		}
 		delete(so.Words, word)
-		return fmt.Sprintf("deleted %s%s", word, Suffix), nil
+		return fmt.Sprintf("deleted %s\n", word), nil
 	}
 	s := strings.Join(noEmpty[1:], " ")
 	so.Words[word] = s
-	return fmt.Sprintf(`defined %s : %s%s`, word, s, Suffix), nil
+	return fmt.Sprintf("defined %s : %s\n", word, s), nil
 }
 
 // parseToken parses token that should be one word and either pushes it to the
@@ -140,14 +152,13 @@ func (so *StackOperator) parseToken(token string) (toPrint string, err error) {
 // executes it accordingly. Returns the string and error from doing what it
 // needs to do.
 func (so *StackOperator) ExecuteToken(token string) (toPrint string, err error) {
-	prefix := "operation error: "
 	p, present := so.Actions.Get(token)
 	if !present {
 		def, present := so.Words[token]
 		if !present {
 			return "", so.notFound(token)
 		}
-        err := so.ParseInput(def)
+		err := so.ParseInput(def)
 		return string(so.PrintBuf), err
 	}
 	action := p.Value
@@ -158,21 +169,21 @@ func (so *StackOperator) ExecuteToken(token string) (toPrint string, err error) 
 		c = 's'
 	}
 	if stkLen < pops {
-		return "", errors.New(fmt.Sprintf("%s'%s' needs %d value%c in stack%s", prefix, token, pops, c, Suffix))
+		return "", errors.New(fmt.Sprintf("operation error: '%s' needs %d value%c in stack\n", token, pops, c))
 	}
-	if stkLen-pops+action.Pushes > cap(so.Stack.Values) {
-		return "", errors.New(fmt.Sprintf("%s'%s' would overflow stack%s", prefix, token, Suffix))
+	if stkLen-pops+action.Pushes > cap(so.Stack.Values) && !so.Stack.Expandable {
+		return "", errors.New(fmt.Sprintf("operation error: '%s' would overflow stack\n", token))
 	}
 	return action.Call(so)
 }
 
-// Fail pushes all `values` to the stack and returns an error containing
+// Fail pushes all values to the stack and returns an error containing
 // `message`. It also prints Stack.Display if the StackOperator is interactive
 func (so *StackOperator) Fail(message string, values ...float64) error {
 	for _, f := range values {
 		so.Stack.Push(f)
 	}
-	return errors.New(fmt.Sprintf("operation error: %s%s", message, Suffix))
+	return errors.New(fmt.Sprintf("operation error: %s\n", message))
 }
 
 // MakePromptFunc sets the StackOperator.prompt value that will execute any
@@ -180,50 +191,88 @@ func (so *StackOperator) Fail(message string, values ...float64) error {
 // it is called. This function returns an error if it could not make the prompt
 // function (should always return nil).
 func (so *StackOperator) MakePromptFunc(format string, fmtChar byte) error {
+	getTopSetup := func(n int) func(*StackOperator) string {
+		return func(so *StackOperator) string {
+			last := make([]string, n)
+			for i := 0; i < n; i++ {
+				p := n - i - 1
+				l := len(so.Stack.Values)
+				if i > l-1 {
+					last[p] = "N"
+				} else {
+					last[p] = fmt.Sprintf("%v", so.Stack.Values[l-i-1])
+				}
+			}
+			return strings.Join(last, " ")
+		}
+	}
 	promptFuncs := make([]func(*StackOperator) string, 0, strings.Count(format, string(fmtChar)))
 	promptFmt := []byte(format)
 	for i := 0; i < len(format)-1; i++ {
 		if format[i] == fmtChar {
 			next := format[i+1]
+			sb := new(strings.Builder)
+			var extra int
+			for j := i + 1; next >= '0' && next <= '9' && j < len(format)-1; j++ {
+				sb.WriteByte(next)
+				next = format[j+1]
+				extra++
+			}
+			conv, err := strconv.Atoi(sb.String())
+			if err != nil {
+				conv = cap(so.Stack.Values)
+			}
+			so.formatters['t'] = getTopSetup(conv)
 			f := so.formatters[next]
 			if f != nil {
 				promptFuncs = append(promptFuncs, f)
 				promptFmt[i] = '%'
 				promptFmt[i+1] = 's'
+				for j := 0; j < extra; j++ {
+					promptFmt[i+j+2] = 0
+				}
 			}
 			if next != fmtChar {
 				i++
 			}
 		}
 	}
-	promptSplit := strings.SplitAfter(string(promptFmt), "%s")
-	if len(promptSplit) != len(promptFuncs)+1 {
-		return errors.New(fmt.Sprintf("Something done gone wrong with the prompt...\nspecifyers: %d, functions: %d", len(promptSplit)-1, len(promptFuncs)))
+	buf := []byte{}
+	for _, b := range promptFmt {
+		if b != 0 {
+			buf = append(buf, b)
+		}
 	}
+	noNull := string(buf)
+	if n := strings.Count(noNull, "%s"); n != len(promptFuncs) {
+		return errors.New(fmt.Sprintf("Prompt screwed up...\nformat num = %d : func num = %d", n, len(promptFuncs)))
+	}
+	vals := make([]any, len(promptFuncs))
 
 	so.Prompt = func() string {
-		sb := new(strings.Builder)
 		for i, f := range promptFuncs {
-			sb.WriteString(fmt.Sprintf(promptSplit[i], f(so)))
+			vals[i] = f(so)
 		}
-		sb.WriteString(promptSplit[len(promptSplit)-1])
-		return sb.String()
+		return fmt.Sprintf(noNull, vals...)
 	}
+
 	return nil
 }
 
 // NewStackOperator returns a pointer to a new StackOperator, initialized to
 // given arguments and a default set of defined words and formatters.
-func NewStackOperator(actions *OrderedMap[string, *Action], maxStack int, interactive bool, strict bool) *StackOperator {
+func NewStackOperator(actions *OrderedMap[string, *Action], maxStack int, interactive bool, noDisplay bool, strict bool) *StackOperator {
 	var displayFmt string
+	displayFmt = "%s\n"
 	if interactive {
-		displayFmt = "[ %s ]" + Suffix
-	} else {
-		displayFmt = "%s" + Suffix
+		displayFmt = "[ %s ]\n"
 	}
-	notFound := func(s string) error { return nil }
+	if noDisplay {
+		displayFmt = ""
+	}
+	notFound := func(string) error { return nil }
 	if strict {
-		notFound = func(s string) error { return errors.New(fmt.Sprintf("command not found: %q\n", s)) }
+		notFound = func(s string) error { return errors.New(fmt.Sprintf("command not found: %s\n", s)) }
 	}
 	stackCap := maxStack
 	if maxStack < 0 {
@@ -234,18 +283,12 @@ func NewStackOperator(actions *OrderedMap[string, *Action], maxStack int, intera
 		Actions:     actions,
 		notFound:    notFound,
 		Interactive: interactive,
-		Words:       map[string]string{},
+		Words:       make(map[string]string),
 		formatters: map[byte]func(*StackOperator) string{
 			'l': func(so *StackOperator) string { return fmt.Sprint(cap(so.Stack.Values)) },
-			't': func(so *StackOperator) string {
-				l := len(so.Stack.Values)
-				if l == 0 {
-					return "NA"
-				}
-				return fmt.Sprint(so.Stack.Values[l-1])
-			},
 			'c': func(so *StackOperator) string { return fmt.Sprint(len(so.Stack.Values)) },
 			's': func(so *StackOperator) string { return fmt.Sprint(so.Stack.Stash) },
+			't': func(*StackOperator) string { return "" },
 		},
 	}
 }